Improved performance and file type support.
I implemented a dictionary lookup as the script was very slow on large projects (2-3 hours with 30-40k images). While the newer script is not as good at telling you why it failed to find a particular image (it can't tell you if there was a duplicate image or if it simply didn't find a match). 

PNG files are also now supported.